+++ src/day_05/task1.rs
@@ -1,6 +1,6 @@
 use crate::day_05::{load_file, process_input, pages_in_correct_order, middle_page};
 
-pub fn task1(file_path: String) -> i32 {
+pub fn task1(file_path: String) -> usize {
     // Load the file content and handle potential errors.
     let data = load_file(&file_path).expect("Failed to read the file.");
 
@@ -11,5 +11,5 @@ pub fn task1(file_path: String) -> i32 {
         } else {
             0
         }
-    })
+    }) as usize
 }
