+++ src/day_05/task2.rs
@@ -1,7 +1,7 @@
 use std::cmp::Ordering;
 use crate::day_05::{load_file, matches_order, pages_in_correct_order, process_input, middle_page};
 
-pub fn task2(file_path: String) -> i32 {
+pub fn task2(file_path: String) -> usize {
     // Load the file content and handle potential errors.
     let data = load_file(&file_path).expect("Failed to read the file.");
 
@@ -13,7 +13,7 @@ pub fn task2(file_path: String) -> i32 {
         } else {
             0
         }
-    })
+    }) as usize
 }
 
 fn sort_pages(pages: &mut Vec<i32>, rules: &[(i32, i32)]) {
