+++ src/day_06/task1.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 use crate::day_06::{read_file, get_guard_position, get_guard_direction, guard_step};
 
-pub fn task1(file_path: String) -> i32 {
+pub fn task1(file_path: String) -> usize {
     let mut field = read_file(file_path).expect("Failed to read the file.");
     let mut guard_position = get_guard_position(&field);
     let mut guard_direction = get_guard_direction(&field, &guard_position);
@@ -18,5 +18,5 @@ pub fn task1(file_path: String) -> i32 {
         &mut visited_positions,
     ) {}
 
-    visited_positions.len() as i32
+    visited_positions.len()
 }
