+++ src/main.rs
@@ -26,7 +26,7 @@ mod day_23;
 mod day_24;
 mod day_25;
 
-fn day_fn(day: i32, task: i32) -> fn(String) -> i32 {
+fn day_fn(day: i32, task: i32) -> fn(String) -> usize {
     match day {
         1 => match task {
             1 => day_01::task1::task1,
